requirements.txt
google-genai
mcp
uvicorn
aiohttp
google-generativeai 
httpx 
python-dotenv 
anthropic
colorlog